LayoutTests/platform/ios/imported/w3c/web-platform-tests/svg/interact/focus-inside-hidden-svg-containers-03-expected.txt

start
middle
end

FAIL Hidden SVG Tab focus test assert_equals: End element should be focused after pressing Tab expected Element node <text tabindex="3" id="end" x="50" y="10">end</text> but got Element node <text tabindex="1" id="start" x="10" y="10">start</text>
